requirements.txt
Flask==3.0.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
requests==2.31.0
cryptography==42.0.5
